backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic>=2.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
groq
hindsight-client
pytest>=8.0.0